# Use pool for pubsub `Publish`'s waitgroups