@LK4D4 2016-02-17
@cpuguy83 can you try with wg per publisher? Seems like `Publish` can't be called concurrently anyway.