@LK4D4 2016-02-17
Another idea is to move wg.Wait out of RLock, that can improve concurrency a lot.