@cpuguy83 2016-02-17
Probably don't need the rlock in general here either... reads are faster everywhere I think.