@GordonTheTurtle 2016-02-17
Job: Docker-PRs-WoW-TP4 [FAILED](https://jenkins.dockerproject.org/job/Docker-PRs-WoW-TP4/1127/console):

~~~console

---
	      56	docker-1.11.0-dev.exe.md5  0%  100%  
	    New File  		      88	docker-1.11.0-dev.exe.sha256  0%  100%  
	    New File  		  27.1 m	docker.exe  0.0%  3.6%  7.3% 11.0% 14.7% 18.3% 22.0% 25.7% 29.4% 33.1% 36.7% 40.4% 44.1% 47.8% 51.5% 55.1% 58.8% 62.5% 66.2% 69.9% 73.5% 77.2% 80.9% 84.6% 88.3% 91.9% 95.6% 99.3%100%  

------------------------------------------------------------------------------

               Total    Copied   Skipped  Mismatch    FAILED    Extras
    Dirs :         1         0         1         0         0         0
   Files :         4         4         0         0         0         0
   Bytes :   54.35 m   54.35 m         0         0         0         0
   Times :   0:00:00   0:00:00                       0:00:00   0:00:00


   Speed :           1017754928 Bytes/sec.
   Speed :            58236.404 MegaBytes/min.
   Ended : Wednesday, February 17, 2016 7:39:10 PM

+ ec=0
+ set +x
INFO: Linking the built binary to /d/CI/CI-dc
---

~~~