@LK4D4 2016-02-17
@cpuguy83 So I think best will be pool + not Locking on Wait. It's not possible to use same wg without locking on wait.