@cpuguy83 2016-02-17
Need the RLock while we wait otherwise can call `Close` and end up closing the underlying channels we're still trying to send on.