@LK4D4 2016-02-17
@cpuguy83 then it worth to try with one wg per publisher.