@cpuguy83 2016-02-17
@LK4D4 In that case, Publish is concurrent and wouldn't work. `RLock` would allow multiple callers of `Publish`.